Adding instrument sub view
