+++ cms/src/web/profiles/open_pension/modules/open_pension_reclamation/src/Entity/InstrumentSubType.php
@@ -34,7 +34,7 @@ use Drupal\user\EntityOwnerTrait;
  *   admin_permission = "access instrument sub type overview",
  *   entity_keys = {
  *     "id" = "id",
- *     "label" = "id",
+ *     "label" = "code",
  *     "uuid" = "uuid",
  *     "owner" = "uid"
  *   },
@@ -49,88 +49,14 @@ use Drupal\user\EntityOwnerTrait;
  */
 class InstrumentSubType extends ContentEntityBase implements InstrumentSubTypeInterface {
 
-  use EntityChangedTrait;
-  use EntityOwnerTrait;
+  use ReclamationEntityFieldsHelper;
 
-  /**
-   * {@inheritdoc}
-   */
-  public function preSave(EntityStorageInterface $storage) {
-    parent::preSave($storage);
-    if (!$this->getOwnerId()) {
-      // If no owner has been set explicitly, make the anonymous user the owner.
-      $this->setOwnerId(0);
-    }
-  }
+  protected static function fieldsMetadata() {
+    $fields = [];
 
-  /**
-   * {@inheritdoc}
-   */
-  public static function baseFieldDefinitions(EntityTypeInterface $entity_type) {
-
-    $fields = parent::baseFieldDefinitions($entity_type);
-
-    $fields['status'] = BaseFieldDefinition::create('boolean')
-      ->setLabel(t('Status'))
-      ->setDefaultValue(TRUE)
-      ->setSetting('on_label', 'Enabled')
-      ->setDisplayOptions('form', [
-        'type' => 'boolean_checkbox',
-        'settings' => [
-          'display_label' => FALSE,
-        ],
-        'weight' => 0,
-      ])
-      ->setDisplayConfigurable('form', TRUE)
-      ->setDisplayOptions('view', [
-        'type' => 'boolean',
-        'label' => 'above',
-        'weight' => 0,
-        'settings' => [
-          'format' => 'enabled-disabled',
-        ],
-      ])
-      ->setDisplayConfigurable('view', TRUE);
-
-    $fields['uid'] = BaseFieldDefinition::create('entity_reference')
-      ->setLabel(t('Author'))
-      ->setSetting('target_type', 'user')
-      ->setDefaultValueCallback(static::class . '::getDefaultEntityOwner')
-      ->setDisplayOptions('form', [
-        'type' => 'entity_reference_autocomplete',
-        'settings' => [
-          'match_operator' => 'CONTAINS',
-          'size' => 60,
-          'placeholder' => '',
-        ],
-        'weight' => 15,
-      ])
-      ->setDisplayConfigurable('form', TRUE)
-      ->setDisplayOptions('view', [
-        'label' => 'above',
-        'type' => 'author',
-        'weight' => 15,
-      ])
-      ->setDisplayConfigurable('view', TRUE);
-
-    $fields['created'] = BaseFieldDefinition::create('created')
-      ->setLabel(t('Authored on'))
-      ->setDescription(t('The time that the instrument sub type was created.'))
-      ->setDisplayOptions('view', [
-        'label' => 'above',
-        'type' => 'timestamp',
-        'weight' => 20,
-      ])
-      ->setDisplayConfigurable('form', TRUE)
-      ->setDisplayOptions('form', [
-        'type' => 'datetime_timestamp',
-        'weight' => 20,
-      ])
-      ->setDisplayConfigurable('view', TRUE);
-
-    $fields['changed'] = BaseFieldDefinition::create('changed')
-      ->setLabel(t('Changed'))
-      ->setDescription(t('The time that the instrument sub type was last edited.'));
+    $fields['code'] = self::simpleTextField(t('Code'));
+    $fields['type'] = self::simpleTextField(t('Type'), FALSE);
+    $fields['stat'] = self::simpleTextField(t('Stat'), FALSE);
 
     return $fields;
   }
